Alternative option for MDMcleaner in case of fail
The default workflow of MDMCleaner raised sometimes exceptions when
checking reference database ambiguity.
The focus here is the identification of contaminated contigs more than
database integrity, therefore the `--fast_run` flag can be used if the
default workflow failed.

The source for the flexible rule is: https://stackoverflow.com/a/73403710